Split Yams into modules and a Core with minimal dependencies
- This change involves splitting Yams into a Core and modules.
The Core contains Yams core functionality and does not depend on specific
cloud platform or storage technology. More specifically, the Azure blob
storage and the distributed locking mechanism have been abstracted away
so that they can be easily interchanged with different implementations
(as an example, take a look at the EndToEndTest which replaces the blob
storage with local file system).

- Several improvements have been made to the update session mechanism
which is used to support upgrade domains such as better error handling and
retry logic.

- I got rid of the annoying consoles that pops-up when running the tests!
You can now run the tests in the background without even noticing them.